### 2024-03-09 #6457 @editorialbot
[PRE REVIEW]: The ARC-OPT Library for Whole-Body Control of Robotic Systems
**Submitting author:** @dmronga (<a href="http://orcid.org/0000-0002-8457-1278">Dennis Mronga</a>)
**Repository:** https://github.com/ARC-OPT/wbc
**Branch with paper.md** (empty if default branch): joss
**Version:** master
