+++ paper.bib
@@ -41,7 +41,8 @@
   journal={2020 IEEE-RAS 20th International Conference on Humanoid Robots (Humanoids)},
   year={2021},
   pages={400-407},
-  url={https://api.semanticscholar.org/CorpusID:231709495}
+  url={https://api.semanticscholar.org/CorpusID:231709495},
+  doi = {10.1109/HUMANOIDS47582.2021.9555770}
 }
 
 
@@ -65,6 +66,7 @@ number    = {4},
 pages     = {3355--3362},
 year      = {2019},
 publisher = {IEEE}
+doi       = {10.1109/LRA.2019.2926664}
 }
 
 @inproceedings{bambade2022,
@@ -78,6 +80,7 @@ publisher = {IEEE}
   PDF = {https://inria.hal.science/hal-03683733/file/Yet_another_QP_solver_for_robotics_and_beyond.pdf},
   HAL_ID = {hal-03683733},
   HAL_VERSION = {v1},
+  doi = {10.15607/rss.2022.xviii.040}
 }
 
 @article{delprete2016,
@@ -96,7 +99,8 @@ publisher = {IEEE}
   PDF = {https://hal.science/hal-01136936v2/file/paper_201403_hal.pdf},
   HAL_ID = {hal-01136936},
   HAL_VERSION = {v2},
-  NOTE = {Software Repository: \url{https://github.com/stack-of-tasks/tsid}}
+  NOTE = {Software Repository: \url{https://github.com/stack-of-tasks/tsid}},
+  doi = {10.1142/s0219843615500449}
 }
 
 @ARTICLE{Ferreau2014,
@@ -107,7 +111,8 @@ publisher = {IEEE}
       volume = {6},
       number = {4},
       pages = {327--363},
-      keywords = {qpOASES, parametric quadratic programming, active set method, model predictive control}
+      keywords = {qpOASES, parametric quadratic programming, active set method, model predictive control},
+      doi = {10.1007/s12532-014-0071-1}
     }
 
 
@@ -156,7 +161,7 @@ volume = {141},
 pages = {103779},
 year = {2021},
 issn = {0921-8890},
-doi = {https://doi.org/10.1016/j.robot.2021.103779},
+doi = {10.1016/j.robot.2021.103779},
 url = {https://www.sciencedirect.com/science/article/pii/S0921889021000646},
 author = {Dennis Mronga and Frank Kirchner},
 keywords = {Context-adaptive control, Constraint-based robot control, Programming-by-demonstration, Gaussian mixture regression, Dual-arm manipulation},
@@ -169,14 +174,16 @@ abstract = {Constraint-based control approaches offer a flexible way to specify 
     year = {2020},
     pages = {1--17},
     journal = {Advanced Robotics},
-    publisher = {Taylor & Francis Online}
+    publisher = {Taylor & Francis Online},
+    doi = {10.1080/01691864.2020.1721322}
 }
 
 @inproceedings{carpentier2019pinocchio,
    title={The Pinocchio C++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives},
    author={Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
    booktitle={IEEE International Symposium on System Integrations (SII)},
-   year={2019}
+   year={2019},
+   doi = {10.1109/sii.2019.8700380}
 }
 
 @Article{Popescu2022,
