+++ paper.bib
@@ -65,7 +65,7 @@ volume    = {4},
 number    = {4},
 pages     = {3355--3362},
 year      = {2019},
-publisher = {IEEE}
+publisher = {IEEE},
 doi       = {10.1109/LRA.2019.2926664}
 }
 
