### 2024-04-15 #6457 @dmronga
@danielskatz  I am clicking on [this](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) link, then I get the following:

![grafik](https://github.com/openjournals/joss-reviews/assets/8993546/9e681c44-24b8-49cb-82a7-4a9895b47dea)

I am asking for access, and that is it. From what I get, somebody should give me to access this file, but that did not happen up to now.

Best,
Dennis
