### 2024-04-15 #6457 @danielskatz
Ah - I'm not sure where you found that, as we don't use it anymore.  See the [first comment](https://github.com/openjournals/joss-reviews/issues/6457#issue-2177134396) in this issue, which points to https://reviewers.joss.theoj.org/lookup
