### 2024-04-22 0e6870f
Make Contact constraints 3-dimensional, this seems to be more stable
