+++ src/constraints/ContactsVelocityConstraint.cpp
@@ -9,12 +9,12 @@ namespace wbc{
         uint nj = robot_model->noOfJoints();
         uint nc = contacts.size();
 
-        A_mtx.resize(nc*6, nj);
-        b_vec.resize(nc*6);
+        A_mtx.resize(nc*3, nj);
+        b_vec.resize(nc*3);
         b_vec.setZero();
 
-        for(int i = 0; i < nc; ++i)
-            A_mtx.block(i*6, 0, 6, nj) = contacts[i].active * robot_model->bodyJacobian(robot_model->worldFrame(), contacts.names[i]);
+        for(uint i = 0; i < nc; ++i)
+            A_mtx.block(i*3, 0, 3, nj) = contacts[i].active * robot_model->bodyJacobian(robot_model->worldFrame(), contacts.names[i]).topRows<3>();
     }
 
 
