+++ src/constraints/ContactsAccelerationConstraint.cpp
@@ -21,7 +21,7 @@ namespace wbc{
         for(uint i = 0; i < contacts.size(); i++){
             const base::Acceleration& a = robot_model->spatialAccelerationBias(robot_model->worldFrame(), contacts.names[i]);
             b_vec.segment(i*3, 3) = -a.linear;
-            A_mtx.block(i*3,  0, 3, nj) = robot_model->bodyJacobian(robot_model->worldFrame(), contacts.names[i]).topRows<3>();
+            A_mtx.block(i*3,  0, 3, nj) = robot_model->spaceJacobian(robot_model->worldFrame(), contacts.names[i]).topRows<3>();
         }
     }
 
