### 2024-04-22 dc398db
Important bug fix in ContactsAccelerationConstraint
