### 2024-04-29 #6696 @editorialbot
[REVIEW]: The ARC-OPT Library for Whole-Body Control of Robotic Systems
**Submitting author:** @dmronga (<a href="http://orcid.org/0000-0002-8457-1278">Dennis Mronga</a>)
**Repository:** https://github.com/ARC-OPT/wbc
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.0
