### 2024-05-03 40173bb
Fix task weighting (Issue #64). Task weights should be multiplied by task matrix and task reference
