### 2024-05-06 79c0e55
Distinguish between active and inactive contacts in constraints
