### 2024-05-06 1f74115
Remove irrelevant variable
