+++ src/robot_models/rbdl/RobotModelRBDL.cpp
@@ -99,7 +99,6 @@ bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
     for(int i = 0; i < actuated_joint_names.size(); i++)
         selection_matrix(i, jointIndex(actuated_joint_names[i])) = 1.0;
 
-    contact_points = cfg.contact_points.names;
     active_contacts = cfg.contact_points;
 
     LOG_DEBUG("------------------- WBC RobotModelRBDL -----------------");
