### 2024-05-06 154389f
Do not check for problem size change in Eiquadprog solver, since the solver deals with this internally
