+++ src/tasks/WrenchForwardTask.cpp
@@ -23,16 +23,15 @@ void WrenchForwardTask::update(RobotModelPtr robot_model){
     ref_frame_pose = robot_model->rigidBodyState(config.root, config.ref_frame).pose;
     ref_frame_rotation = ref_frame_pose.orientation.toRotationMatrix();
     weights_root.segment(0,3) = ref_frame_rotation * weights.segment(0,3);
-    weights_root.segment(3,3) = ref_frame_rotation * weights.segment(3,3);
-    weights_root = weights_root.cwiseAbs();
+    weights_root = weights;//weights_root.cwiseAbs();
 }
 
 void WrenchForwardTask::setReference(const base::samples::RigidBodyStateSE3& ref){
 
-    if(!ref.hasValidWrench()){
-        LOG_ERROR("Task %s has invalid force and/or torque", config.name.c_str())
-        throw std::invalid_argument("Invalid task reference value");
-    }
+//    if(!ref.hasValidWrench()){
+//        LOG_ERROR("Task %s has invalid force and/or torque", config.name.c_str())
+//        throw std::invalid_argument("Invalid task reference value");
+//    }
 
     if(ref.time.isNull())
         this->time = base::Time::now();
@@ -40,7 +39,6 @@ void WrenchForwardTask::setReference(const base::samples::RigidBodyStateSE3& ref
         this->time = ref.time;
 
     this->y_ref.segment(0,3) = ref.wrench.force;
-    this->y_ref.segment(3,3) = ref.wrench.torque;
 }
 
 
