+++ src/constraints/RigidbodyDynamicsConstraint.cpp
@@ -10,7 +10,7 @@ namespace wbc{
         uint na = robot_model->noOfActuatedJoints();
         uint nc = contacts.size();
 
-        uint nv = reduced ? (nj + nc*6) : (nj + na + nc*6);
+        uint nv = reduced ? (nj + nc*3) : (nj + na + nc*3);
 
         if(reduced) // no torques in qp, consider only floating base dynamics
         {
@@ -22,8 +22,8 @@ namespace wbc{
             A_mtx.block(0,  0, 6, nj) =  robot_model->jointSpaceInertiaMatrix().topRows<6>();
             for(uint i = 0; i < contacts.size(); i++){
                 if(contacts[i].active)
-                    A_mtx.block(0, nj+i*6, 6, 6) = -robot_model->bodyJacobian(robot_model->worldFrame(), contacts.names[i]).transpose().topRows<6>();
-            b_vec = -robot_model->biasForces().topRows<6>();
+                    A_mtx.block(0, nj+i*3, 6, 3) = -robot_model->spaceJacobian(robot_model->worldFrame(), contacts.names[i]).topRows<3>().transpose().topRows<6>();
+                b_vec = -robot_model->biasForces().topRows<6>();
             }
         }
         else
@@ -37,7 +37,7 @@ namespace wbc{
             A_mtx.block(0, nj, nj, na) = -robot_model->selectionMatrix().transpose();
             for(uint i = 0; i < contacts.size(); i++){
                 if(contacts[i].active)
-                    A_mtx.block(0, nj+na+i*6, nj, 6) = -robot_model->bodyJacobian(robot_model->worldFrame(), contacts.names[i]).transpose();
+                    A_mtx.block(0, nj+na+i*3, nj, 3) = -robot_model->spaceJacobian(robot_model->worldFrame(), contacts.names[i]).topRows(3).transpose();
                 b_vec = -robot_model->biasForces();
             }
         }

[…]
