### 2024-05-23 9606e52
RobotModelPinocchio: Expect floating base velocity/acceleration in hybrid coordinates (meaning angular part is expressed locally aligned to world)
