### 2024-05-28 ARC-OPT/wbc#94 @mhubii
JOSS: State of the field
Please always advise if incorrect.

Re: State of the field

Should other libraries such as etasl or optas be mentioned or is the WBC domain most relevant?

Refers https://github.com/openjournals/joss-reviews/issues/6696
