### 2024-05-28 ARC-OPT/wbc#95 @mhubii
JOSS: OSQP
Please always advise if incorrect.

Is there a specific reason why `osqp` was not considered a solver? How difficult would the inclusion of `osqp` be?

I am referring to this benchmark: https://github.com/qpsolvers/qpbenchmark#plots

Refers https://github.com/openjournals/joss-reviews/issues/6696
