### 2024-05-28 ARC-OPT/wbc#96 @mhubii
JOSS: Minor documentation
Please always advise if incorrect.

Published topics, subscribed topics etc empty, is this intended, maybe consider removing these headers? https://arc-opt.github.io/Documentation/tutorials/ros2_introduction.html

Links broken: https://arc-opt.github.io/Documentation/tutorials/ros2_nullspace_control.html

Refers https://github.com/openjournals/joss-reviews/issues/6696
