### 2024-05-28 ARC-OPT/wbc#97 @mhubii
JOSS: Docker
Could you please provide me a Dockerfile to test the library?

Refers: https://github.com/openjournals/joss-reviews/issues/6696
