### 2024-05-28 ARC-OPT/wbc#98 @mhubii
JOSS: Paper - Example
Please always advise if incorrect.

The example section of the paper states an optimization problem. Is it correct, that the joint torques and external forces are part of the optimization? From my best understanding, this problem seeks to minimize ddq subject to constraints. It is stated that tau and f are minimized, too.

Refers https://github.com/openjournals/joss-reviews/issues/6696
