### 2024-05-28 ARC-OPT/wbc#99 @mhubii
JOSS: Dependencies
Please always advise if incorrect.

In the installation, I can find plenty cloning. Would it make sense to add theses as git submodules in a `third_party` folder?

I am talking about these:

https://github.com/ARC-OPT/wbc/blob/154389f28d5a64c18eb89cd7e1a5281a4dbd55ce/scripts/full_install.sh#L78

Refers https://github.com/openjournals/joss-reviews/issues/6696
