### 2024-05-28 ARC-OPT/wbc#100 @mhubii
JOSS: Tests
Please always advise if incorrect.

First of all, I think there are plenty tests, which is great. I have a minor suggestion, but this might be personal preference.

The tests are somehow run via a shell script: https://github.com/ARC-OPT/wbc/blob/master/scripts/run_tests.sh

Is this common? I have only ever used gtest before and cmake for example provides ways to run these. I am afraid that having a shell script for that might miss some tests should further additions be made.

Refers https://github.com/openjournals/joss-reviews/issues/6696
