### 2024-05-29 18cc05a
Remove QP printouts
