+++ paper.bib
@@ -200,3 +200,25 @@ ISSN = {1424-8220},
 ABSTRACT = {Regardless of recent advances, humanoid robots still face significant difficulties in performing locomotion tasks. Among the key challenges that must be addressed to achieve robust bipedal locomotion are dynamically consistent motion planning, feedback control, and state estimation of such complex systems. In this paper, we investigate the use of an external motion capture system to provide state feedback to an online whole-body controller. We present experimental results with the humanoid robot RH5 performing two different whole-body motions: squatting with both feet in contact with the ground and balancing on one leg. We compare the execution of these motions using state feedback from (i) an external motion tracking system and (ii) an internal state estimator based on inertial measurement unit (IMU), forward kinematics, and contact sensing. It is shown that state-of-the-art motion capture systems can be successfully used in the high-frequency feedback control loop of humanoid robots, providing an alternative in cases where state estimation is not reliable.},
 DOI = {10.3390/s22249853}
 }
+
+@Inbook{Smits2009,
+author="Smits, Ruben
+and De Laet, Tinne
+and Claes, Kasper
+and Bruyninckx, Herman
+and De Schutter, Joris",
+editor="Hahn, Hernsoo
+and Ko, Hanseok
+and Lee, Sukhan",
+title="iTASC: A Tool for Multi-Sensor Integration in Robot Manipulation",
+bookTitle="Multisensor Fusion and Integration for Intelligent Systems: An Edition of  the Selected Papers from the IEEE International Conference on Multisensor Fusion and Integration for Intelligent Systems 2008",
+year="2009",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="235--254",
+abstract="iTASC (acronym for `instantaneous task specification using constraints) deschutter2006 is a systematic constraint-based approach to specify complex tasks of general sensor-based robot systems. iTASC integrates both instantaneous task specification and estimation of geometric uncertainty in a unified framework. Automatic derivation of controller and estimator equations follows from a geometric task model that is obtained using a systematic task modeling procedure. The approach applies to a large variety of robot systems (mobile robots, multiple robot systems, dynamic human-robot interaction, etc.), various sensor systems, and different robot tasks. Using an example task, this paper shows that iTASC is a powerful tool for multi-sensor integration in robot manipulation. The example task includes multiple sensors: encoders, a force sensor, cameras, a laser distance sensor and a laser scanner. The paper details the systematic modeling procedure for the example task and elaborates on the task specific choice of two types of task coordinates: feature coordinates, defined with respect to object and feature frames, which facilitate the task specification, and uncertainty coordinates to model geometric uncertainty. Experimental results for the example task are presented.",
+isbn="978-3-540-89859-7",
+doi="10.1007/978-3-540-89859-7_17",
+url="https://doi.org/10.1007/978-3-540-89859-7_17"
+}
+
