### 2024-05-30 ARC-OPT/wbc#94 @dmronga
Good point. I added itasc (https://ieeexplore.ieee.org/document/4648032), which is I think the basis of etasl. Etasl itself is used for high-level task specification, so I would rather not mention it. Regarding optas, it seems that it is used for MPC and trajectory optimization. This library is only for WBC (instantaneous optimization).
