### 2024-05-30 ARC-OPT/wbc#95 @dmronga
Yes, integrating osqp is definitely on my TODO list. I will integrate it
