### 2024-05-30 ARC-OPT/wbc#96 @dmronga
Fixed broken links and documentation
