### 2024-05-30 ARC-OPT/wbc#98 @dmronga
Yes it is correct. [This](https://github.com/ARC-OPT/wbc/blob/master/src/scenes/acceleration_reduced_tsid/AccelerationSceneReducedTSID.hpp) controller solves for joint accelerations and contact forces, while [this](https://github.com/ARC-OPT/wbc/blob/master/src/scenes/acceleration_tsid/AccelerationSceneTSID.hpp) controller solves for joint accelerations , torques and contact forces
