### 2024-06-03 a6553d2
Fix wrong matrix dimension in AccelerationReducedTSID::solve()
