### 2024-06-05 87bafe5
Add possibility to online update the URDF model
