### 2024-06-05 6fe10f6
Add osqp solver
