### 2024-06-05 01aeebf
Remove printouts in test case
