### 2024-06-05 89766be
Add tests in cmake
