### 2024-06-06 ARC-OPT/wbc#100 @dmronga
Fixed by #102
