### 2024-06-06 ARC-OPT/wbc#95 @dmronga
Fixed by #101
