### 2024-06-07 b4b0931
Adapt AccelerationSceneTSID according to AccelerationSceneReducedTSID
