### 2024-06-07 aa59d6f
Remove KDL robot model, it lacks too many features compared to RBDL and pinocchio
