+++ src/robot_models/kdl/RobotModelKDL.cpp
@@ -1,515 +1,0 @@
-#include "RobotModelKDL.hpp"
-#include "KinematicChainKDL.hpp"
-#include <kdl_parser/kdl_parser.hpp>
-#include <base-logging/Logging.hpp>
-#include "../../core/RobotModelConfig.hpp"
-#include <kdl/treeidsolver_recursive_newton_euler.hpp>
-#include <algorithm>
-#include <tools/URDFTools.hpp>
-#include <kdl/chaindynparam.hpp>
-
-namespace wbc{
-
-RobotModelRegistry<RobotModelKDL> RobotModelKDL::reg("kdl");
-
-RobotModelKDL::RobotModelKDL(){
-}
-
-RobotModelKDL::~RobotModelKDL(){
-}
-
-void RobotModelKDL::clear(){
-
-    RobotModel::clear();
-
-    full_tree = KDL::Tree();
-    kdl_chain_map.clear();
-}
-
-bool RobotModelKDL::configure(const RobotModelConfig& cfg){
-
-    clear();
-
-    // 1. Load Robot Model
-
-    robot_model_config = cfg;
-    robot_urdf = loadRobotURDF(cfg.file_or_string);
-    if(!robot_urdf){
-        LOG_ERROR("Unable to parse urdf model")
-        return false;
-    }
-    URDFTools::applyJointBlacklist(robot_urdf, cfg.joint_blacklist);
-    base_frame =  robot_urdf->getRoot()->name;
-
-    // Joint names from URDF without floating base
-    std::vector<std::string> joint_names_urdf = URDFTools::jointNamesFromURDF(robot_urdf);
-
-    // Add floating base
-    has_floating_base = cfg.floating_base;
-    world_frame = base_frame;
-    if(cfg.floating_base){
-        joint_names_floating_base = URDFTools::addFloatingBaseToURDF(robot_urdf);
-        world_frame = robot_urdf->getRoot()->name;
-    }
-
-    // Read Joint Limits
-    URDFTools::jointLimitsFromURDF(robot_urdf, joint_limits);
-
-    // Parse KDL Tree
-    if(!kdl_parser::treeFromUrdfModel(*robot_urdf, full_tree)){
-        LOG_ERROR("Unable to load KDL Tree");
-        return false;
-    }
-
-    joint_names = independent_joint_names = joint_names_floating_base + joint_names_urdf;
-    actuated_joint_names = joint_names_urdf;
-    joint_state.elements.resize(joint_names.size());
-    joint_state.names = joint_names;
-
-    // 2. Verify consistency of URDF and config
-
-    // All contact point have to be a valid link in the robot URDF
-    for(auto c : cfg.contact_points.names){
-        if(!hasLink(c)){
-            LOG_ERROR("Contact point %s is not a valid link in the robot model", c.c_str());
-            return false;
-        }
-    }
-
-    // 3. Create data structures
-
-    q.resize(noOfJoints());
-    qd.resize(noOfJoints());
-    qdd.resize(noOfJoints());
-    tau.resize(noOfJoints());
-    zero.resize(noOfJoints());
-    zero.data.setZero();
-    active_contacts = cfg.contact_points;
-    joint_space_inertia_mat.resize(noOfJoints(), noOfJoints());
-    bias_forces.resize(noOfJoints());
-    selection_matrix.resize(noOfActuatedJoints(),noOfJoints());
-    selection_matrix.setZero();
-    for(int i = 0; i < actuated_joint_names.size(); i++)
-        selection_matrix(i, jointIndex(actuated_joint_names[i])) = 1.0;
-
-    for(const auto &it : full_tree.getSegments()){
[…]
