+++ scripts/full_install.sh
@@ -20,7 +20,7 @@ make -j8 && sudo make install && cd ../..
 # URDF
 sudo apt-get -y install liburdfdom-headers-dev liburdfdom-dev 
 
-# Clone WBC repo to have the patches for KDL and qpOASES
+# Clone WBC repo to have the patches 
 git clone https://github.com/ARC-OPT/wbc.git
 
 # RBDL
@@ -31,29 +31,12 @@ mkdir build && cd build
 cmake .. -DRBDL_BUILD_ADDON_URDFREADER=ON
 make -j8 && sudo make install && cd ../..
 
-# KDL 
-git clone https://github.com/orocos/orocos_kinematics_dynamics.git -b v1.5.1
-mkdir orocos_kinematics_dynamics/orocos_kdl/build
-cd orocos_kinematics_dynamics/orocos_kdl/build
-cmake .. 
-make -j8 && sudo make install &&  cd ../../..
-
 # Pinocchio
 git clone --branch v2.6.8 --recurse-submodules https://github.com/stack-of-tasks/pinocchio.git
 cd pinocchio
 mkdir build && cd build
 cmake .. -DBUILD_PYTHON_INTERFACE=OFF -DBUILD_UNIT_TESTS=OFF 
 make -j8 && sudo make install && cd ../..
-
-# KDL parser
-sudo apt-get -y install libtinyxml2-dev
-git clone https://github.com/ros/kdl_parser.git -b 1.14.1
-cd kdl_parser/kdl_parser
-mkdir patches && cp ../../wbc/patches/kdl_parser.patch patches
-git apply patches/kdl_parser.patch
-mkdir build && cd build 
-cmake .. 
-make -j8 && sudo make install && cd ../../..
 
 # If not done yet, setup a ssh key pair using the command `ssh-keygen` and add the 
 # key from `~/.ssh/id_rsa.pub `to the keys in your Gitlab account.
@@ -103,7 +86,7 @@ make -j8 && sudo make install && cd ../..
 
 # WBC
 mkdir wbc/build && cd wbc/build
-cmake .. -DROBOT_MODEL_RBDL=ON -DROBOT_MODEL_KDL=ON -DSOLVER_PROXQP=ON -DSOLVER_EIQUADPROG=ON -DSOLVER_QPSWIFT=ON -DSOLVER_OSQP=ON -DCMAKE_BUILD_TYPE=RELEASE
+cmake .. -DROBOT_MODEL_RBDL=ON -DSOLVER_PROXQP=ON -DSOLVER_EIQUADPROG=ON -DSOLVER_QPSWIFT=ON -DSOLVER_OSQP=ON -DCMAKE_BUILD_TYPE=RELEASE
 make -j8 && sudo make install && cd ..
 
 sudo ldconfig

[…]
