### 2024-06-09 e60daa8
Allow configuration of interpolator wrt. whether it should use internal (estimated) state or current joint state as a starting point
