### 2024-06-10 6c5a914
Remove rh5v2 tutorial for now, as it seems broken
