### 2024-07-18 b327a0d
Refactoring: - Remove dependency to base-types,base-logging, introduce native types - Restructure Scene configuration - Use world as default base frame - Remove deprecated scenes, solvers and robot models -
