+++ src/robot_models/rbdl/RobotModelRBDL.cpp
@@ -2,7 +2,7 @@
 #include "tools/URDFTools.hpp"
 #include <rbdl/rbdl_utils.h>
 #include <rbdl/addons/urdfreader/urdfreader.h>
-#include <base-logging/Logging.hpp>
+#include "../tools/Logger.hpp"
 
 using namespace RigidBodyDynamics;
 
@@ -23,10 +23,6 @@ void RobotModelRBDL::clear(){
     rbdl_model = std::make_shared<Model>();
 }
 
-void RobotModelRBDL::updateFloatingBase(const base::samples::RigidBodyStateSE3& floating_base_state_in){
-
-}
-
 bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
 
     clear();
@@ -36,7 +32,7 @@ bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
     robot_model_config = cfg;
     robot_urdf = loadRobotURDF(cfg.file_or_string);
     if(!robot_urdf){
-        LOG_ERROR("Unable to parse robot urdf model");
+        log(logERROR)<<"Unable to parse urdf model";
         return false;
     }
     base_frame = robot_urdf->getRoot()->name;
@@ -44,7 +40,7 @@ bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
     joint_names = URDFTools::jointNamesFromURDF(robot_urdf);
 
     // Temporary workaround: RBDL does not support rotations of the link invertias. Set them to zero.
-    // Note: This will lead to inconsistent behavior in acceleration-based WBC
+    // Note: This will lead to inconsistent behavior in acceleration-Eigend WBC
     for(auto &l : robot_urdf->links_){
         if(l.second->inertial)
             l.second->inertial->origin.rotation.setFromRPY(0,0,0);
@@ -54,277 +50,216 @@ bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
     doc->SaveFile(robot_urdf_file);
 
     if(!Addons::URDFReadFromFile(robot_urdf_file.c_str(), rbdl_model.get(), cfg.floating_base)){
-        LOG_ERROR_S << "Unable to parse urdf from file " << robot_urdf_file << std::endl;
+        log(logERROR) << "Unable to parse urdf from file " << robot_urdf_file;
         return false;
     }
 
-    // Add floating base to robot_urdf. We will not load this URDF model in RBDL, since RBDL adds its own floating base.
+    // Add floating Eigen to robot_urdf. We will not load this URDF model in RBDL, since RBDL adds its own floating Eigen.
     // However, we need the robot_urdf for some internal functionalities.
     has_floating_base = cfg.floating_base;
     world_frame = base_frame;
-    if(cfg.floating_base){
-        joint_names_floating_base = URDFTools::addFloatingBaseToURDF(robot_urdf);
-        world_frame = robot_urdf->getRoot()->name;
-    }
+    if(cfg.floating_base)
+        world_frame = "world";
 
     actuated_joint_names = joint_names;
-    joint_names = independent_joint_names = joint_names_floating_base + joint_names;
+    independent_joint_names = joint_names;
 
     // Read Joint Limits
-    URDFTools::jointLimitsFromURDF(robot_urdf, joint_limits);
+    URDFTools::jointLimitsFromURDF(robot_urdf, joint_limits, joint_names);
 
-    // 2. Verify consistency of URDF and config
+    // 2. Create data structures
 
-    // All contact point have to be a valid link in the robot URDF
-    for(auto c : cfg.contact_points.names){
-        if(!hasLink(c)){
-            LOG_ERROR("Contact point %s is not a valid link in the robot model", c.c_str());
+    joint_state.resize(joint_names.size());
+
[…]
