+++ src/tasks/JointAccelerationTask.hpp
@@ -1,15 +1,18 @@
-#ifndef JOINTACCELERATIONTASK_HPP
-#define JOINTACCELERATIONTASK_HPP
+#ifndef WBC_TASKS_JOINTACCELERATIONTASK_HPP
+#define WBC_TASKS_JOINTACCELERATIONTASK_HPP
 
-#include "JointTask.hpp"
+#include "../core/Task.hpp"
 
 namespace wbc{
 /**
  * @brief Implementation of a Joint velocity task.
  */
-class JointAccelerationTask : public JointTask{
+class JointAccelerationTask : public Task{
+    std::vector<std::string> joint_names;
 public:
-    JointAccelerationTask(TaskConfig config, uint n_robot_joints);
+    JointAccelerationTask(TaskConfig config,
+                          const std::vector<std::string>& joint_names,
+                          uint nj);
     virtual ~JointAccelerationTask() = default;
 
     /**
@@ -24,7 +27,10 @@ public:
      * @param ref Joint reference input. Vector size has ot be same number of task variables. Joint Names have to match the task joint names.
      * Each entry has to have a valid acceleration. All other entries will be ignored.
      */
-    virtual void setReference(const base::commands::Joints& ref);
+    void setReference(const Eigen::VectorXd& ref);
+
+    /** @brief Returns the vector of joint names used by the task*/
+    const std::vector<std::string>& jointNames(){return joint_names;}
 };
 
 using JointAccelerationTaskPtr = std::shared_ptr<JointAccelerationTask>;

[…]
