### 2024-07-25 ARC-OPT/wbc#97 @mhubii
one more thing, could you just provide me the dockerfile as opposed to the container
