### 2024-07-30 2fb42b7
Update robot models, to do tze required computations only once per cycle
