+++ src/core/RobotModel.hpp
@@ -41,16 +41,22 @@ protected:
     types::JointState joint_state;
     Eigen::MatrixXd joint_space_inertia_mat;
     Eigen::MatrixXd com_jac;
-    Eigen::MatrixXd selection_matrix;
     Eigen::VectorXd bias_forces;
-    types::SpatialAcceleration spatial_acc_bias;
-    types::RigidBodyState com_rbs;
-    types::RigidBodyState rbs;
-    Eigen::VectorXd q, qd, qdd, tau, zero_jnt;
     types::SpatialAcceleration zero_acc;
-    Eigen::MatrixXd space_jac;
-    Eigen::MatrixXd body_jac;
+    Eigen::MatrixXd selection_matrix;
+    types::RigidBodyState com_rbs;
+    Eigen::VectorXd q, qd, qdd, tau, zero_jnt;
     bool configured,updated;
+    std::map<std::string,Eigen::MatrixXd> space_jac_map;
+    std::map<std::string,Eigen::MatrixXd> body_jac_map;
+    std::map<std::string,types::Pose> pose_map;
+    std::map<std::string,types::Twist> twist_map;
+    std::map<std::string,types::SpatialAcceleration> acc_map;
+    std::map<std::string,types::SpatialAcceleration> spatial_acc_bias_map;
+    bool compute_inertia_mat;
+    bool compute_com;
+    bool compute_com_jac;
+    bool compute_bias_forces;
 
 public:
     RobotModel();
@@ -113,15 +119,15 @@ public:
 
     /** Returns the pose of the body defined by frame_id. Pose will be computed with respect to world frame (floating base robot) or
      *  robot root link (fixed base)*/
-    virtual const types::Pose &pose(const std::string &frame_id) = 0;
+    virtual const types::Pose &pose(const std::string &frame_id, const bool recompute = false) = 0;
 
     /** Returns the twist of the body defined by frame_id. Twist will be in "local-world-aligned" (hybrid) representation, i.e., with respect to a frame
       * attached to the given body, but aligned to world coordinates (floating base robot) or robot root link (fixed base)*/
-    virtual const types::Twist &twist(const std::string &frame_id) = 0;
+    virtual const types::Twist &twist(const std::string &frame_id, const bool recompute = false) = 0;
 
     /** Returns the spatial acceleration of the body defined by frame_id. Twist will be in "local-world-aligned" (hybrid) representation, i.e., with respect to a frame
       * attached to the given body, but aligned to world coordinates (floating base robot) or robot root link (fixed base)*/
-    virtual const types::SpatialAcceleration &acceleration(const std::string &frame_id) = 0;
+    virtual const types::SpatialAcceleration &acceleration(const std::string &frame_id, const bool recompute = false) = 0;
 
 
     /** @brief Returns the Space Jacobian for the given frame. The order of the Jacobian's columns will be the same as in the model (alphabetic). Note that
@@ -130,32 +136,32 @@ public:
       * @param frame_id Tip frame of the Jacobian. Has to be a valid link in the robot model.
       * @return A 6 x nj matrix, where nj is the number of joints + number of floating coordinates, i.e. 6 in case of a floating base robot.
       */
-    virtual const Eigen::MatrixXd &spaceJacobian(const std::string &frame_id) = 0;
+    virtual const Eigen::MatrixXd &spaceJacobian(const std::string &frame_id, const bool recompute = false) = 0;
 
     /** @brief Returns the Body Jacobian for the given frame. The order of the Jacobian's columns will be the same as in the model (alphabetic).
       * @param frame_id Reference frame of the Jacobian. Has to be a valid link in the robot model.
       * @return A 6 x nj matrix, where nj is the number of joints + number of floating coordinates, i.e. 6 in case of a floating base robot.
       */
[…]
