### 2024-07-30 4196c19
Bug fix in wrench task
