+++ tutorials/rh5v2/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_executable(rh5v2 rh5v2.cpp)
 target_link_libraries(rh5v2
                       wbc-solvers-qpoases
-                      wbc-scenes-acceleration_tsid
+                      wbc-scenes-acceleration_reduced_tsid
                       wbc-controllers
                       wbc-robot_models-pinocchio)
 
@@ -9,7 +9,7 @@ if(ROBOT_MODEL_HYRODYN)
    add_executable(rh5v2_hybrid rh5v2_hybrid.cpp)
    target_link_libraries(rh5v2_hybrid
                          wbc-solvers-qpoases
-                         wbc-scenes-acceleration_tsid
+                         wbc-scenes-acceleration_reduced_tsid
                          wbc-controllers
                          wbc-robot_models-hyrodyn)
 endif()                         
