### 2024-08-08 fc09320
Pass robot model to tasks for simplicity of use
