+++ src/scenes/acceleration_tsid/AccelerationSceneTSID.cpp
@@ -93,7 +93,7 @@ const HierarchicalQP& AccelerationSceneTSID::update(){
                task->type == com_acceleration ||
                task->type == wrench_forward);
 
-        task->update(robot_model);
+        task->update();
 
         // If the activation value is zero, also set reference to zero. Activation is usually used to switch between different
         // task phases and we don't want to store the "old" reference value, in case we switch on the task again
