+++ src/tasks/CartesianVelocityTask.cpp
@@ -3,13 +3,13 @@
 namespace wbc{
 
 CartesianVelocityTask::CartesianVelocityTask(TaskConfig config,
+                                             RobotModelPtr robot_model,
                                              const std::string &tip_frame,
-                                             const std::string &ref_frame,
-                                             uint nj)
-    : Task(config, 6, nj, TaskType::spatial_velocity), tip_frame(tip_frame), ref_frame(ref_frame){
+                                             const std::string &ref_frame)
+    : Task(config, robot_model, 6, TaskType::spatial_velocity), tip_frame(tip_frame), ref_frame(ref_frame){
 }
 
-void CartesianVelocityTask::update(RobotModelPtr robot_model){
+void CartesianVelocityTask::update(){
     
     // Task Jacobian
     A = robot_model->spaceJacobian(tip_frame);
