+++ src/scenes/acceleration_tsid/test/test_acceleration_scene_tsid.cpp
@@ -52,9 +52,9 @@ BOOST_AUTO_TEST_CASE(simple_test){
     // Configure scene
     CartesianAccelerationTaskPtr cart_task;
     cart_task = make_shared<CartesianAccelerationTask>(TaskConfig("cart_pos_ctrl",0,{1,1,1,1,1,1},1),
+                                                       robot_model,
                                                        "RH5_Root_Link",
-                                                       "world",
-                                                       robot_model->nj());
+                                                       "world");
     AccelerationSceneTSID wbc_scene(robot_model, solver, 1e-3);
     BOOST_CHECK_EQUAL(wbc_scene.configure({cart_task}), true);
 
