### 2024-08-08 ca7bf1a
Update solvers: Expose parameters
