+++ src/core/RobotModel.hpp
@@ -26,6 +26,39 @@ namespace wbc{
 class RobotModel{
 protected:
     void clear();
+    void resetData();
+    void updateData();
+
+    struct Pose{
+        Pose(){needs_recompute = true;}
+        bool needs_recompute;
+        types::Pose data;
+    };
+    struct Twist{
+        Twist(){needs_recompute = true;}
+        bool needs_recompute;
+        types::Twist data;
+    };
+    struct SpatialAcceleration{
+        SpatialAcceleration(){needs_recompute = true;}
+        bool needs_recompute;
+        types::SpatialAcceleration data;
+    };
+    struct Matrix{
+        Matrix(){needs_recompute = true;}
+        bool needs_recompute;
+        Eigen::MatrixXd data;
+    };
+    struct Vector{
+        Vector(){needs_recompute = true;}
+        bool needs_recompute;
+        Eigen::VectorXd data;
+    };
+    struct RigidBodyState{
+        RigidBodyState(){needs_recompute = true;}
+        bool needs_recompute;
+        types::RigidBodyState data;
+    };
 
     std::vector<Contact> contacts;
     Eigen::Vector3d gravity;
@@ -39,24 +72,23 @@ protected:
     urdf::ModelInterfaceSharedPtr robot_urdf;
     bool has_floating_base;
     types::JointState joint_state;
-    Eigen::MatrixXd joint_space_inertia_mat;
-    Eigen::MatrixXd com_jac;
-    Eigen::VectorXd bias_forces;
-    types::SpatialAcceleration zero_acc;
     Eigen::MatrixXd selection_matrix;
-    types::RigidBodyState com_rbs;
     Eigen::VectorXd q, qd, qdd, tau, zero_jnt;
+
+    std::vector<Matrix> com_jac;
+    std::vector<Matrix> joint_space_inertia_mat;
+    std::vector<Vector> bias_forces;
+    std::vector<RigidBodyState> com_rbs;
+    std::map<std::string,Matrix> space_jac_map;
+    std::map<std::string,Matrix> body_jac_map;
+    std::map<std::string,Pose> pose_map;
+    std::map<std::string,Twist> twist_map;
+    std::map<std::string,SpatialAcceleration> acc_map;
+    std::map<std::string,SpatialAcceleration> spatial_acc_bias_map;
+    bool fk_needs_recompute, fk_with_zero_acc_needs_recompute;
+
+    types::SpatialAcceleration zero_acc;
     bool configured,updated;
-    std::map<std::string,Eigen::MatrixXd> space_jac_map;
-    std::map<std::string,Eigen::MatrixXd> body_jac_map;
-    std::map<std::string,types::Pose> pose_map;
-    std::map<std::string,types::Twist> twist_map;
-    std::map<std::string,types::SpatialAcceleration> acc_map;
-    std::map<std::string,types::SpatialAcceleration> spatial_acc_bias_map;
-    bool compute_inertia_mat;
-    bool compute_com;
-    bool compute_com_jac;
-    bool compute_bias_forces;
 
 public:
     RobotModel();
@@ -119,15 +151,15 @@ public:
 
     /** Returns the pose of the body defined by frame_id. Pose will be computed with respect to world frame (floating base robot) or
      *  robot root link (fixed base)*/
-    virtual const types::Pose &pose(const std::string &frame_id, const bool recompute = false) = 0;
+    virtual const types::Pose &pose(const std::string &frame_id) = 0;
 
     /** Returns the twist of the body defined by frame_id. Twist will be in "local-world-aligned" (hybrid) representation, i.e., with respect to a frame
       * attached to the given body, but aligned to world coordinates (floating base robot) or robot root link (fixed base)*/
[…]
