### 2024-08-12 7b37a14
Fix options in proxQP and qpSwift
