### 2024-08-12 520352d
Remove cout
