### 2024-08-16 ARC-OPT/wbc#97 @dmronga
Sorry, I missed your query. Here is the docker file and install script: https://github.com/ARC-OPT/wbc/tree/master/docker. Both have to be in the same folder. Build/run with
docker build -t image_name .
docker run -i -t image_name
