### 2024-08-16 863a8d7
Add acados interface as a solver
