+++ src/solvers/CMakeLists.txt
@@ -11,3 +11,6 @@ endif()
 if(SOLVER_OSQP)
     add_subdirectory(osqp)
 endif()
+if(SOLVER_ACADOS)
+    add_subdirectory(acados)
+endif()
