### 2024-08-16 ec996d1
Free memory in acados solver
