### 2024-08-19 d447663
Rename acados solver to hpipm solver
