+++ src/solvers/CMakeLists.txt
@@ -11,6 +11,6 @@ endif()
 if(SOLVER_OSQP)
     add_subdirectory(osqp)
 endif()
-if(SOLVER_ACADOS)
-    add_subdirectory(acados)
+if(SOLVER_HPIPM)
+    add_subdirectory(hpipm)
 endif()
