### 2024-08-19 5e305d4
Add missing registry entry for OSQP solver
