### 2024-08-19 b678614
Rephrase Contact friction cone constraint as it failed on HPIPM solver
