### 2024-08-23 f4ea8cc
Fix in HPIPM solver
