### 2024-08-27 ARC-OPT/wbc#97 @mhubii
awesome, thank you for sharing. This should get me up to speed.

On a related note, would it be possible to merge the `RUN` commands for caching, e.g. in

https://github.com/ARC-OPT/wbc/blob/069dee2be489f2154e3d4892868b218fceaf4e9f/docker/Dockerfile#L3

```docker
RUN apt-get update -y && \
    apt-get upgrade -y && \
    apt-get -y install git cmake build-essential libboost-system-dev libboost-program-options-dev libboost-thread-dev libboost-test-dev pkg-config libeigen3-dev libboost-filesystem-dev
```

Please feel free to close this issue after
