### 2024-09-04 f5ed186
Attempt to fix a warm start issue in the solvers. Previously the solver checked the QP size and reinitialized the QP ( and performed cold start) when the QP size changes. However, there may be situations where the QP size remaines the same, but warm start is infeasible, e.g., in quadruped trotting, when the contacts changes from FL,RR to FR,RL.
