+++ src/scenes/acceleration_reduced_tsid/AccelerationSceneReducedTSID.hpp
@@ -44,6 +44,17 @@ protected:
     Eigen::VectorXd robot_acc, solver_output_acc;
     std::vector<types::Wrench> contact_wrenches;
     double hessian_regularizer;
+    std::vector<Contact> contacts;
+
+    bool contactsHaveChanged(const std::vector<Contact>& old_contacts, const std::vector<Contact>& new_contacts){
+        if(old_contacts.size() != new_contacts.size())
+            return true;
+        for(uint i = 0; i < old_contacts.size(); i++){
+            if(old_contacts[i].active != new_contacts[i].active)
+                return true;
+        }
+        return false;
+    }
 
 public:
     AccelerationSceneReducedTSID(RobotModelPtr robot_model, QPSolverPtr solver, const double dt);
