### 2024-09-04 b4ddb0a
Reduce regularization term for the sake of stability
