### 2024-09-05 65e6485
Use warm_start = 1 (only primal solutions) for HPIPM solver
