### 2024-09-09 ARC-OPT/wbc#107 @mhubii
JOSS: Link to doc
Under about, could you please add the documentation as website? See below

![about](https://github.com/user-attachments/assets/d0bf75dc-3bd5-4506-ac8f-8c7303ab73ba)

![website](https://github.com/user-attachments/assets/5e60f205-1e41-4594-a70f-89e366859175)

If you could do the same for `wbc_ros` / `wbc_msgs`, would be great.

On this topic, in the tutorials, e.g. https://arc-opt.github.io/Documentation/tutorials/vel_introductory_example.html, how does one jump back to the landing page? Can you add a navigation bar easily?

Refers to https://github.com/openjournals/joss-reviews/issues/6696
