### 2024-09-09 ARC-OPT/wbc#108 @mhubii
JOSS: Docker installation
Under https://arc-opt.github.io/Documentation/installation/installation_no_rock.html, would it be possible to explain how to install / run docker, along the lines of `docker build . -t wbc` or what not

Refers to https://github.com/openjournals/joss-reviews/issues/6696
