### 2024-09-09 ARC-OPT/wbc#109 @mhubii
JOSS: Links broken
1. Documentation: https://arc-opt.github.io/Documentation/tutorials/vel_introductory_example.html

    "as shown **here**", link broken

2. ROS: https://github.com/ARC-OPT/wbc_ros#unit-tests

    "The unit tests for the wbc library can be found **here**.", link broken

Refers to https://github.com/openjournals/joss-reviews/issues/6696
