### 2024-09-09 ARC-OPT/wbc#110 @mhubii
JOSS: typos
- `QPOases` -> `qpOASES`, e.g. here: https://arc-opt.github.io/Documentation/tutorials/vel_using_different_solver.html
- Also minor: "allows inequality constraints, e.g.., joint velocity limits": e.g.. -> e.g. (https://arc-opt.github.io/Documentation/tutorials/vel_using_different_solver.html)
- ROS2 -> ROS 2 (https://shouldiputaspacebetweenrosand2.com/) :D

Refers to https://github.com/openjournals/joss-reviews/issues/6696
