### 2024-09-09 ARC-OPT/wbc#111 @mhubii
JOSS: Mute solver
I've been running through introduction https://arc-opt.github.io/Documentation/tutorials/vel_introductory_example.html inside Docker.

Reading through Doxygen (https://arc-opt.github.io/wbc/classwbc_1_1QPSolver.html), I can't find a means of "muting" the output. Muting output seems? desirable

Refers to https://github.com/openjournals/joss-reviews/issues/6696
