### 2024-09-09 ARC-OPT/wbc#112 @mhubii
JOSS: Doxygen
Could you please add a direct link to Doxygen on https://arc-opt.github.io/Documentation/

Am I missing something? This seems a little hidden away.

Refers to https://github.com/openjournals/joss-reviews/issues/6696
