### 2024-09-09 ARC-OPT/wbc#113 @mhubii
JOSS: Running tutorials
While the tutorials run fine, and I got experience with C++, it isn't quite obvious how the tutorials are run. Am I missing something?

E.g. here https://arc-opt.github.io/Documentation/tutorials/vel_introductory_example.html

Could you please add

```shell
cd build/tutorials/kuka_iiwa
./cart_pos_ctrl_hls_hierarchies
```

or similar.

Refers to https://github.com/openjournals/joss-reviews/issues/6696
