### 2024-09-09 #6696 @mhubii
so I have done a second round of reviews, targeting functionality and performance. The library works as intended and except for some suggestions, I am happy. I will now have to investigate its integration into ROS 2 in more detail
