### 2024-09-09 #6696 @mhubii
I think the ROS 2 support has a little cleaning up to do. This depends a little on what the main focus of this submission is. Arguably, the WBC library in itself already provides a lot of functionality with the ROS integration being the cherry on top..
