### 2024-09-10 ARC-OPT/wbc#107 @dmronga
Added the link in the about sections.

Navigation bar is not so easy, but I added links at the bottom of each tutorial to navigate through the tutorials more easily
