### 2024-09-10 ARC-OPT/wbc#109 @dmronga
Fixed/removed broken links
