### 2024-09-10 ARC-OPT/wbc#110 @dmronga
Fixed minor
