### 2024-09-10 ARC-OPT/wbc#111 @dmronga
The output is not in the solver but in the tutorial itself to show the user what happens. The solver itself should not write output on the cmd line.
