### 2024-09-10 ARC-OPT/wbc#108 @mhubii
using `git clone git@github.com:ARC-OPT/wbc.git` does require access rights, consider https
