### 2024-09-12 #6696 @dmronga
@mhubii No the wbc_py lib is outdated. The original publication should only consider the wbc library. But it is definitely good that you provided remarks on the ros2 interface.
