### 2024-09-12 #6696 @mhubii
Okay sounds good. With the cleaned dependencies, I'll do one more run for the ROS integration, otherwise, I am happy with the submission and think it adds a valuable contribution to JOSS.

One more remark on adding robot descriptions to a controller package. I get the design choice and the ease of use, but there is no clear separation of concerns. But since this is no issue in itself, I'd leave it to the author for future work.
