### 2024-11-18 #6696 @adi3
@JHartzer @sea-bass thank you for stepping up! you can generate your respective checklists by issuing `@editorialbot generate my checklist`. More details [here](https://github.com/openjournals/joss-reviews/issues/6696#issue-2269574925). Please let me know if I can assist with anything!
