### 2024-11-20 ARC-OPT/wbc#114 @sea-bass
[JOSS Review] Re-enable `rh5v2` examples or document that they currently do not work
While testing this package, I noticed that the `rh5v2` examples are commented out from the `CMakeLists.txt` with a comment saying they are broken.

Ideally this would be fixed before a submission, but if not possible please add comments in the documentation that these are currently not working. If the maintainers are too busy to resolve this, I would recommend creating an issue with some guidelines for contributors to help out.

https://github.com/openjournals/joss-reviews/issues/6696
