### 2024-11-20 ARC-OPT/wbc#115 @sea-bass
[JOSS Review] Support Ubuntu 24.04 / ROS 2 Jazzy
Right now, the tool is only documented to work with Ubuntu 20.04 and 22.04, but 24.04 has been out for a while.

My personal laptop runs 24.04, so I could only test in the Docker container which limited what I could try out.

https://github.com/openjournals/joss-reviews/issues/6696
