### 2024-11-20 ARC-OPT/wbc#116 @sea-bass
[JOSS Review] How do users reproduce the tutorials' visualization?
The introductory tutorials show videos of a simulated KUKA iiwa and some plots that help explain the concepts.

However, running the examples I noticed there was no visualization except for numbers on the console.

I would strongly suggest adding visualization instructions to the documentation, or even a more general section showing how things can be better visualized.

https://github.com/openjournals/joss-reviews/issues/6696
