### 2024-11-20 ARC-OPT/wbc#117 @sea-bass
[JOSS Review] Paper Feedback
This issue tracks my comments on the JOSS submission publication for ARC-OPT WBC.

Please refer to https://github.com/openjournals/joss-reviews/issues/6696 for my overall comments and links to other issues.

---

- [x] Overall, the paper is very short. While JOSS publications are not intended to be as rigorous as academic conference/journals, I think more details throughout the whole publications are necessary.
- [x] At least one more figure is necessary that shows visualization of a robot model from tje actual software -- something like the KUKA iiwa videos embedded in the tutorial documentation, but of course it would be more impactful if you picked a more complex kinematic tree such as a full humanoid.
- [x] The "State of the field" criterion for the paper is not currently being met, and needs to be added. Specifically, there needs to be comparisons with other open-source tools for whole-body control. Please consider doing some review including tools like these, among others that I am for sure missing.
  - [x] As users of Pinocchio, you can refer to related tools like [Humanoid Path Planner (HPP)](https://humanoid-path-planner.github.io/hpp-doc/), [OCS2](https://leggedrobotics.github.io/ocs2/) and [Crocoddyl](https://github.com/loco-3d/crocoddyl).
  - [x] Other popular stack of tasks approaches include Stephane Caron's [Pink](https://github.com/stephane-caron/pink), also based on Pinocchio, and the related [Mink](https://github.com/kevinzakka/mink) from Kevin Zakka, which is based on MuJoCo.
  - [x] I've also recently learned about the [humanoid control stack from IIT](https://github.com/ami-iit). These are only a few suggestions, and I'm sure many more can be found.
- [x] It is typical, and in my opinion required, for a JOSS paper to speak somewhat about the software.
  - [x] The design is briefly introduced, but could use more detail (as with the whole paper)
  - [x] The Example introduced how to set up a QP for whole-body control generically, but I strongly suggest then showing an actual code snippet from WBC showing how easy it is to translate the mathematical formulation into code. It's a major point of submissions to this journal.
  - [x] Finally, you are doing your great work a disservice by not mentioning that these tools have Python bindings and ROS 2 support, and integrate with `ros2_control`. Please include this information, as it's good!
- [x] There is one sentence that (very) briefly explains WBC's support for floating base dynamics, friction cone constraints, cost weighting, and task hierarchies. I think this part in particular could use a bit more fleshing out, as well as pointing to the fact that most of these are actually available as tutorials. The only one I didn't find was the friction cone constraint, but everything else was for sure there.
