### 2024-11-20 #6696 @sea-bass
Hello @dmronga -- thank you for this great work. It was a pleasure to review it, especially as a fellow user of Pinocchio that is currently starting to learn more about optimal planning and control.

Overall, the software is in great shape and all the examples ran (except for the `rh5v2` which was commented out). However, I was on Ubuntu 24.04 and while I could use the Docker to try core WBC, I didn't have an easy way to try the ROS 2 support. Most importantly, I was expecting the tutorials to have some visualization beyond console output, which I think will be an important addition.

The paper is quite minimal and needs some more substance put into it. Most of this involves better communicating all the functionality that is already present in the WBC stack. You can find more details in https://github.com/ARC-OPT/wbc/issues/117

Besides that, I've written down some other issues and will break them down into must have and nice to have.

**Must have** (required)
* https://github.com/ARC-OPT/wbc/issues/114
* https://github.com/ARC-OPT/wbc/issues/116
* https://github.com/ARC-OPT/wbc_ros/issues/13
* https://github.com/ARC-OPT/Documentation/issues/3

**Nice to have** (not required)
* https://github.com/ARC-OPT/wbc/issues/115

Please feel free to discuss individual things in the respective issues, and I look forward to seeing revisions. This is really great work, and just needs a little bit of push on the tutorials and publication. Thank you!
