### 2024-11-26 ARC-OPT/wbc#117 @dmronga
Thanks a lot for the detailed comments, I will consider them as much as possible. Just some rebuttals from my side:
* Paper length: Agreed I will add more detailed explanations, but according to [here](https://joss.readthedocs.io/en/latest/paper.html) there is a page limit of 1000 words.
* State of field should be covered by the statement of need section. I mention some related frameworks there, but I agree that it is very short and I will extend this section. The frameworks that you mention are, however, not really comparable to ARC-OPT, which considers reactive (instantaneous) whole-body motion control:
   - HPP is for motion planning
   - OCS2 and Crocoddyl deal with optimal control, mostly in the sense of trajectory optimization, although also online implementation are available
   - PINK/MINK are IK solvers, where ARC-OPT is also intended for dynamics computations
   - The humanoid control stack uses TSID, which is mentioned in my paper

Still I will have another look around what is out and relate the existing frameworks to my software.
