### 2024-11-27 ARC-OPT/wbc#117 @sea-bass
Sounds good. As you said, I don't think that many extra words are needed, but some more content that better describes the specifics of your tool's capabilities and comparison to other related OSS software.

The initial paper, to me, mostly described the underlying theory but not the implementation.
