+++ paper.bib
@@ -17,6 +17,34 @@
   urldate = {2021-08-02}
 }
 
+@online{controlit2021,
+  author = {Human-Centered Robotics Lab of the University of Texas at Austin},
+  title = {ControlIt! - A Whole Body Operational Space Control Middleware},
+  year = 2021,
+  url = {https://github.com/liangfok/controlit},
+  urldate = {2021-07-29}
+}
+
+@misc{drake2021,
+ author = "Russ Tedrake and the Drake Development Team",
+ title = "Drake: Model-based design and verification for robotics",
+ year = 2019,
+ url = "https://drake.mit.edu"
+}
+
+@article{Feng2015,
+abstract = {One popular approach to controlling humanoid robots is through inverse kinematics (IK) with stiff joint position tracking. On the other hand, inverse dynamics (ID) based approaches have gained increasing acceptance by providing compliant motions and robustness to external perturbations. However, the performance of such methods is heavily dependent on high quality dynamic models, which are often very difficult to produce for a physical robot. IK approaches only require kinematic models, which are much easier to generate in practice. In this paper, we supplement our previous work with ID-based controllers by adding IK, which helps compensate for modeling errors. The proposed full body controller is applied to three tasks in the DARPA Robotics Challenge (DRC) Trials in Dec. 2013.},
+author = {Feng, Siyuan and Whitman, Eric and Xinjilefu, X. and Atkeson, Christopher G.},
+doi = {10.1109/HUMANOIDS.2014.7041347},
+isbn = {9781479971749},
+issn = {21640580},
+journal = {IEEE-RAS International Conference on Humanoid Robots},
+pages = {120--127},
+publisher = {IEEE},
+title = {{Optimization based full body control for the atlas robot}},
+year = {2015}
+}
+
 @article{2019_Kumar_HyRoDynApproach_IDETC,
     author = {Kumar, Shivesh and Szadkowski, Kai Alexander von and Mueller, Andreas and Kirchner, Frank},
     title = {An Analytical and Modular Software Workbench for Solving Kinematics and Dynamics of Series-Parallel Hybrid Robots},
