+++ paper.bib
@@ -25,6 +25,15 @@
   urldate = {2021-07-29}
 }
 
+@software{pink2024,
+  title = {{Pink: Python inverse kinematics based on Pinocchio}},
+  author = {Caron, Stéphane and De Mont-Marin, Yann and Budhiraja, Rohan and Bang, Seung Hyeon and Domrachev, Ivan and Nedelchev, Simeon},
+  license = {Apache-2.0},
+  url = {https://github.com/stephane-caron/pink},
+  version = {3.1.0},
+  year = {2024}
+}
+
 @misc{drake2021,
  author = "Russ Tedrake and the Drake Development Team",
  title = "Drake: Model-based design and verification for robotics",
