+++ paper.bib
@@ -41,6 +41,13 @@
  url = "https://drake.mit.edu"
 }
 
+@misc{ros2control2024,
+ author = "ros2_control Development Team",
+ title = "ros2_control documentation",
+ year = 2024,
+ url = "https://control.ros.org/humble/"
+}
+
 @article{Feng2015,
 abstract = {One popular approach to controlling humanoid robots is through inverse kinematics (IK) with stiff joint position tracking. On the other hand, inverse dynamics (ID) based approaches have gained increasing acceptance by providing compliant motions and robustness to external perturbations. However, the performance of such methods is heavily dependent on high quality dynamic models, which are often very difficult to produce for a physical robot. IK approaches only require kinematic models, which are much easier to generate in practice. In this paper, we supplement our previous work with ID-based controllers by adding IK, which helps compensate for modeling errors. The proposed full body controller is applied to three tasks in the DARPA Robotics Challenge (DRC) Trials in Dec. 2013.},
 author = {Feng, Siyuan and Whitman, Eric and Xinjilefu, X. and Atkeson, Christopher G.},
