### 2024-11-29 043660b
Update gains in tutorial
