### 2024-11-29 ARC-OPT/wbc#117 @dmronga
@sea-bass I update the paper draft according to your suggestions. In particular I ...
* Added more explanation on the design
* Added a code snipped showing a simple WBC setup
* Added some screenshots showing the result of the code snippet.
* Extended the state of the Art. I added reference to MINK and further WBC frameworks. The other tools you mentioned imo. serve other purposes. They are either for motion planning or optimal control
* Added a small section on ROS 2 support and Python bindings

Best regards
