### 2024-11-29 ARC-OPT/wbc#117 @sea-bass
Looks great -- thank you!

Re: those other tools serving a different purpose. That's right, and my intention was to have some comparison of tools that solve QPs or nonlinear programs in general for various aspects of planning and control. But I see you've mentioned tools like Drake now as well, which is good.
