+++ tutorials/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectory(kuka_iiwa)
 add_subdirectory(rh5)
-#add_subdirectory(rh5v2)
+add_subdirectory(rh5v2)
