### 2024-11-30 ARC-OPT/wbc#114 @dmronga
I  fixed the rh5v2 tutorial, on velocity level, however. Instead I added another tutorial to show acceleration-level WBC
