### 2024-12-03 1fab3c6
Fix to work on tinyxml2
