+++ src/tools/CMakeLists.txt
@@ -6,13 +6,13 @@ file(GLOB SOURCES RELATIVE ${PROJECT_SOURCE_DIR}/src/types "*.cpp")
 pkg_search_module(base-types REQUIRED IMPORTED_TARGET base-types)
 pkg_search_module(base-logging REQUIRED IMPORTED_TARGET base-logging)
 pkg_search_module(urdfdom REQUIRED IMPORTED_TARGET urdfdom)
-pkg_search_module(tinyxml REQUIRED IMPORTED_TARGET tinyxml)
+pkg_search_module(tinyxml2 REQUIRED IMPORTED_TARGET tinyxml2)
 pkg_search_module(eigen3 REQUIRED IMPORTED_TARGET eigen3)
 
 list(APPEND PKGCONFIG_REQUIRES base-types)
 list(APPEND PKGCONFIG_REQUIRES base-logging)
 list(APPEND PKGCONFIG_REQUIRES urdfdom)
-list(APPEND PKGCONFIG_REQUIRES tinyxml)
+list(APPEND PKGCONFIG_REQUIRES tinyxml2)
 list(APPEND PKGCONFIG_REQUIRES eigen3)
 string (REPLACE ";" " " PKGCONFIG_REQUIRES "${PKGCONFIG_REQUIRES}")
 
@@ -26,7 +26,7 @@ target_link_libraries(${TARGET_NAME} PUBLIC
                       PkgConfig::base-types
                       PkgConfig::base-logging
                       PkgConfig::urdfdom
-                      PkgConfig::tinyxml
+                      PkgConfig::tinyxml2
                       PkgConfig::eigen3)
 
 set_target_properties(${TARGET_NAME} PROPERTIES
