### 2024-12-03 c72bfa4
Update build_and_test_ubuntu24.04.yml
