### 2024-12-03 40fdc91
Fix install script
