### 2024-12-03 a6b18be
Remove direct dependency to tinyxml
