+++ src/robot_models/rbdl/RobotModelRBDL.cpp
@@ -23,10 +23,6 @@ void RobotModelRBDL::clear(){
     rbdl_model = std::make_shared<Model>();
 }
 
-void RobotModelRBDL::updateFloatingBase(const base::samples::RigidBodyStateSE3& floating_base_state_in){
-
-}
-
 bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
 
     clear();
@@ -49,9 +45,9 @@ bool RobotModelRBDL::configure(const RobotModelConfig& cfg){
         if(l.second->inertial)
             l.second->inertial->origin.rotation.setFromRPY(0,0,0);
     }
-    TiXmlDocument *doc = urdf::exportURDF(robot_urdf);
+    auto *doc = urdf::exportURDF(robot_urdf);
     std::string robot_urdf_file = "/tmp/robot.urdf";
-    doc->SaveFile(robot_urdf_file);
+    doc->SaveFile(robot_urdf_file.c_str());
 
     if(!Addons::URDFReadFromFile(robot_urdf_file.c_str(), rbdl_model.get(), cfg.floating_base)){
         LOG_ERROR_S << "Unable to parse urdf from file " << robot_urdf_file << std::endl;
