### 2024-12-04 6daa976
Compatibility to urdfdom 4.0.0
