+++ scripts/full_install.sh
@@ -19,6 +19,8 @@ make -j8 && sudo make install && cd ../..
 
 # URDF
 sudo apt-get -y install liburdfdom-headers-dev liburdfdom-dev 
+# compatability with urdfdom >= 4.0.0 (Ubuntu24.04)´
+sudo apt-get -y install libtinyxml2-dev
 
 # Clone WBC repo to have the patches 
 git clone https://github.com/ARC-OPT/wbc.git
