### 2024-12-05 ARC-OPT/wbc#115 @dmronga
I added support for Ubuntu24.04 and ROS 2 jazzy
