### 2024-12-06 e86cfa2
Add method to get internal pinocchio model
