### 2024-12-07 #6696 @sea-bass
@dmronga has addressed all my review items, and I've finished checking the missing boxes in my review.

I am happy to recommend this for publication in JOSS 🚀

---

One last nitpicky/editorial comment: The paper mentions the "IHCM Whole-Body Controller", but this should be "IHMC".
