### 2024-12-10 #6696 @dmronga
> @dmronga Looks like everything is set. I need one final formality to be done from your side. Please make a tagged release of the project and archive it. Then report the version number and archive DOI here. I'll need that for my post-review checklist. Thanks!

@adi3 The tagged release is called JOSS Paper (tag version 0.3.0). The DOI is 10.5281/zenodo.14353780. Thank you so much for organizing this review!
