### 2024-12-12 8edc5ca
Fix tinyxml2 issue
