+++ tutorials/kuka_iiwa/cart_pos_ctrl_weights.cpp
@@ -76,6 +76,7 @@ int main(){
     ref_pose.orientation.setIdentity();
     pose.position.setZero();
     pose.orientation.setIdentity();
+    ref_twist.setZero();
 
     // Run control loop
     double loop_time = 0.01; // seconds
