### 2024-12-16 #6696 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
