+++ paper.bib
@@ -223,7 +223,7 @@ abstract = {Constraint-based control approaches offer a flexible way to specify 
 }
 
 @inproceedings{carpentier2019pinocchio,
-   title={The Pinocchio C++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives},
+   title={The {P}inocchio {C}++ library -- A fast and flexible implementation of rigid body dynamics algorithms and their analytical derivatives},
    author={Carpentier, Justin and Saurel, Guilhem and Buondonno, Gabriele and Mirabel, Joseph and Lamiraux, Florent and Stasse, Olivier and Mansard, Nicolas},
    booktitle={IEEE International Symposium on System Integrations (SII)},
    year={2019},
